Closes #65
