+++ src/cli.rs
@@ -13,11 +13,13 @@
 // limitations under the License.
 use crate::chain_spec;
 use crate::service;
+use futures::{future, sync::oneshot, Future};
+use log::info;
+use std::cell::RefCell;
 use std::ops::Deref;
 use substrate_cli as cli;
 pub use substrate_cli::{error, IntoExit, NoCustom, VersionInfo};
 use substrate_service::{Roles as ServiceRoles, ServiceFactory};
-use tokio::prelude::Future;
 use tokio::runtime::Runtime;
 
 /// The chain specification option.
@@ -137,4 +139,29 @@ where
 	let _ = runtime.shutdown_on_idle().wait();
 
 	Ok(())
+}
+
+// handles ctrl-c
+pub struct Exit;
+
+impl IntoExit for Exit {
+	type Exit = future::MapErr<oneshot::Receiver<()>, fn(oneshot::Canceled) -> ()>;
+	fn into_exit(self) -> Self::Exit {
+		// can't use signal directly here because CtrlC takes only `Fn`.
+		let (exit_send, exit) = oneshot::channel();
+
+		let exit_send_cell = RefCell::new(Some(exit_send));
+		ctrlc::set_handler(move || {
+			if let Some(exit_send) = exit_send_cell
+				.try_borrow_mut()
+				.expect("signal handler not reentrant; qed")
+				.take()
+			{
+				exit_send.send(()).expect("Error sending exit notification");
+			}
+		})
+		.expect("Error setting Ctrl-C handler");
+
+		exit.map_err(drop)
+	}
 }
